statusok handle

+++ src/api/user_api.rs
@@ -4,7 +4,10 @@ use poem_openapi::{param::Path, payload::Json, OpenApi};
 use crate::db::{self, Db, User};
 use poem::{web::Data, Result};
 
-use super::guards::{AdminGuard, Login};
+use super::{
+    guards::{AdminGuard, Login},
+    StatusOk,
+};
 
 #[OpenApi(prefix_path = "/api/users", tag = "super::Groups::Users")]
 impl UserApi {
@@ -15,15 +18,28 @@ impl UserApi {
     }
     #[oai(path = "/", method = "post")]
     ///add a new user
-    async fn add_user(&self, usr: Json<db::User>, _u: AdminGuard, db: Data<&Db>) -> Result<()> {
+    async fn add_user(
+        &self,
+        usr: Json<db::User>,
+        _u: AdminGuard,
+        db: Data<&Db>,
+    ) -> Result<Json<StatusOk>> {
         db.new_user((*usr).clone()).await?;
-        Ok(())
+        Ok(Json(().into()))
     }
     #[oai(path = "/:id", method = "delete")]
     ///delete a user
-    async fn delete_user(&self, id: Path<String>, db: Data<&Db>, _u: AdminGuard) -> Result<()> {
+    async fn delete_user(
+        &self,
+        id: Path<String>,
+        db: Data<&Db>,
+        u: AdminGuard,
+    ) -> Result<Json<StatusOk>> {
+        if *id == u.id {
+            return Err(crate::Error::CannotDeleteSelf.into());
+        }
         db.delete_user(&id).await?;
-        Ok(())
+        Ok(Json(().into()))
     }
     #[oai(path = "/login", method = "post")]
     ///login using basic auth
